+++ samples/D20Tek.FullSample.Wasm/Program.cs
@@ -10,7 +10,8 @@ builder.RootComponents.Add<HeadOutlet>("head::after");
 
 builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) })
                 .AddMarkdownRenderer()
-                .AddMessageBox();
+                .AddMessageBox()
+                .AddToast();
 
 builder.Services.AddScoped<RegistrationService>();