+++ tests/D20Tek.BlazorComponents.UnitTests/GlobalUsings.cs
@@ -1,5 +1,6 @@
 ﻿global using Bunit;
 global using D20Tek.Functional;
+global using Microsoft.AspNetCore.Components;
 global using Microsoft.AspNetCore.Components.Forms;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.VisualStudio.TestTools.UnitTesting;